fixing bug on first call of a standing

+++ src/app/football/football-resolver.service.ts
@@ -6,12 +6,13 @@ export class FootballResolverService implements Resolve<any> {
   constructor() {}
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
     const previousDate = new Date(JSON.parse(localStorage.getItem('taskCreatedAt') || '[]'));
-    const currentDate = new Date();
-    const differenceInTime = currentDate.getTime() - previousDate.getTime();
-    const differenceInDays = differenceInTime / (1000 * 3600 * 24);
-    console.log(differenceInDays);
-    if (differenceInDays > 1) {
-      localStorage.clear();
+    if(previousDate) {
+      const currentDate = new Date();
+      const differenceInTime = currentDate.getTime() - previousDate.getTime();
+      const differenceInDays = differenceInTime / (1000 * 3600 * 24);
+      if (differenceInDays > 1) {
+        localStorage.clear();
+      }
     }
   }
 }
